requirements.prod.txt
Click==6.7
requests==2.18.4
python-string-utils==0.6.0
